changefeedccl: remove old webhook and pubsub sink code
This PR removes the old webhook and pubsub implementations, having the v2 files
take their place.

Release note: The changefeed.new_webhook_sink_enabled and
changefeed.new_pubsub_sink_enabled settings have been removed as they are now
always enabled.
