+++ pkg/ccl/changefeedccl/BUILD.bazel
@@ -33,10 +33,8 @@ go_library(
         "sink_external_connection.go",
         "sink_kafka.go",
         "sink_pubsub.go",
-        "sink_pubsub_v2.go",
         "sink_sql.go",
         "sink_webhook.go",
-        "sink_webhook_v2.go",
         "telemetry.go",
         "testing_knobs.go",
         "tls.go",
