+++ pkg/ccl/changefeedccl/sink_webhook.go
@@ -1,4 +1,4 @@
-// Copyright 2021 The Cockroach Authors.
+// Copyright 2023 The Cockroach Authors.
 //
 // Licensed as a CockroachDB Enterprise file under the Cockroach Community
 // License (the "License"); you may not use this file except in compliance with
@@ -13,9 +13,7 @@ import (
 	"context"
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/json"
 	"fmt"
-	"hash/crc32"
 	"io"
 	"net"
 	"net/http"
@@ -24,265 +22,67 @@ import (
 	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/changefeedbase"
-	"github.com/cockroachdb/cockroach/pkg/ccl/changefeedccl/kvevent"
-	"github.com/cockroachdb/cockroach/pkg/util/ctxgroup"
-	"github.com/cockroachdb/cockroach/pkg/util/hlc"
+	"github.com/cockroachdb/cockroach/pkg/util/admission"
 	"github.com/cockroachdb/cockroach/pkg/util/httputil"
-	"github.com/cockroachdb/cockroach/pkg/util/retry"
-	"github.com/cockroachdb/cockroach/pkg/util/system"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/errors"
 )
 
-type deprecatedWebhookSink struct {
-	// Webhook configuration.
-	parallelism int
-	retryCfg    retry.Options
-	batchCfg    batchConfig
-	ts          timeutil.TimeSource
-	format      changefeedbase.FormatType
+const (
+	applicationTypeJSON = `application/json`
+	applicationTypeCSV  = `text/csv`
+	authorizationHeader = `Authorization`
+)
 
-	// Webhook destination.
+func isWebhookSink(u *url.URL) bool {
+	switch u.Scheme {
+	// allow HTTP here but throw an error later to make it clear HTTPS is required
+	case changefeedbase.SinkSchemeWebhookHTTP, changefeedbase.SinkSchemeWebhookHTTPS:
+		return true
+	default:
+		return false
+	}
+}
+
+type webhookSinkClient struct {
+	ctx        context.Context
+	format     changefeedbase.FormatType
 	url        sinkURL
 	authHeader string
+	batchCfg   sinkBatchConfig
 	client     *httputil.Client
-
-	// messages are written onto batch channel
-	// which batches matches based on batching configuration.
-	batchChan chan webhookMessage
-
-	// flushDone channel signaled when flushing completes.
-	flushDone chan struct{}
-
-	// errChan is written to indicate an error while sending message.
-	errChan chan error
-
-	// parallelism workers are created and controlled by the workerGroup, running with workerCtx.
-	// each worker gets its own events channel.
-	workerCtx   context.Context
-	workerGroup ctxgroup.Group
-	exitWorkers func() // Signaled to shut down all workers.
-	eventsChans []chan []deprecatedMessagePayload
-	metrics     metricsRecorder
 }
 
-func (s *deprecatedWebhookSink) getConcreteType() sinkType {
-	return sinkTypeWebhook
-}
+var _ SinkClient = (*webhookSinkClient)(nil)
+var _ SinkPayload = (*http.Request)(nil)
 
-type webhookSinkPayload struct {
-	Payload []json.RawMessage `json:"payload"`
-	Length  int               `json:"length"`
-}
-
-type encodedPayload struct {
-	data     []byte
-	alloc    kvevent.Alloc
-	emitTime time.Time
-	mvcc     hlc.Timestamp
-}
-
-func encodePayloadJSONWebhook(messages []deprecatedMessagePayload) (encodedPayload, error) {
-	result := encodedPayload{
-		emitTime: timeutil.Now(),
-	}
-
-	payload := make([]json.RawMessage, len(messages))
-	for i, m := range messages {
-		result.alloc.Merge(&m.alloc)
-		payload[i] = m.val
-		if m.emitTime.Before(result.emitTime) {
-			result.emitTime = m.emitTime
-		}
-		if result.mvcc.IsEmpty() || m.mvcc.Less(result.mvcc) {
-			result.mvcc = m.mvcc
-		}
-	}
-
-	body := &webhookSinkPayload{
-		Payload: payload,
-		Length:  len(payload),
-	}
-	j, err := json.Marshal(body)
-	if err != nil {
-		return encodedPayload{}, err
-	}
-	result.data = j
-	return result, err
-}
-
-func encodePayloadCSVWebhook(messages []deprecatedMessagePayload) (encodedPayload, error) {
-	result := encodedPayload{
-		emitTime: timeutil.Now(),
-	}
-
-	var mergedMsgs []byte
-	for _, m := range messages {
-		result.alloc.Merge(&m.alloc)
-		mergedMsgs = append(mergedMsgs, m.val...)
-		if m.emitTime.Before(result.emitTime) {
-			result.emitTime = m.emitTime
-		}
-		if result.mvcc.IsEmpty() || m.mvcc.Less(result.mvcc) {
-			result.mvcc = m.mvcc
-		}
-	}
-
-	result.data = mergedMsgs
-	return result, nil
-}
-
-type deprecatedMessagePayload struct {
-	// Payload message fields.
-	key      []byte
-	val      []byte
-	alloc    kvevent.Alloc
-	emitTime time.Time
-	mvcc     hlc.Timestamp
-}
-
-// webhookMessage contains either messagePayload or a flush request.
-type webhookMessage struct {
-	flushDone *chan struct{}
-	payload   deprecatedMessagePayload
-}
-
-type batch struct {
-	buffer      []deprecatedMessagePayload
-	bufferBytes int
-}
-
-func (b *batch) addToBuffer(m deprecatedMessagePayload) {
-	b.bufferBytes += len(m.val)
-	b.buffer = append(b.buffer, m)
-}
-
-func (b *batch) reset() {
-	b.buffer = b.buffer[:0]
-	b.bufferBytes = 0
-}
-
-type batchConfig struct {
-	Bytes, Messages int          `json:",omitempty"`
-	Frequency       jsonDuration `json:",omitempty"`
-}
-
-// wrapper structs to unmarshal json, retry.Options will be the actual config
-type retryConfig struct {
-	Max     jsonMaxRetries `json:",omitempty"`
-	Backoff jsonDuration   `json:",omitempty"`
-}
-
-// proper JSON schema for webhook sink config:
-//
-//	{
-//	  "Flush": {
-//		   "Messages":  ...,
-//		   "Bytes":     ...,
-//		   "Frequency": ...,
-//	  },
-//		 "Retry": {
-//		   "Max":     ...,
-//		   "Backoff": ...,
-//	  }
-//	}
-type webhookSinkConfig struct {
-	Flush batchConfig `json:",omitempty"`
-	Retry retryConfig `json:",omitempty"`
-}
-
-func (s *deprecatedWebhookSink) getWebhookSinkConfig(
-	jsonStr changefeedbase.SinkSpecificJSONConfig,
-) (batchCfg batchConfig, retryCfg retry.Options, err error) {
-	retryCfg = defaultRetryConfig()
-
-	var cfg webhookSinkConfig
-	cfg.Retry.Max = jsonMaxRetries(retryCfg.MaxRetries)
-	cfg.Retry.Backoff = jsonDuration(retryCfg.InitialBackoff)
-	if jsonStr != `` {
-		// set retry defaults to be overridden if included in JSON
-		if err = json.Unmarshal([]byte(jsonStr), &cfg); err != nil {
-			return batchCfg, retryCfg, errors.Wrapf(err, "error unmarshalling json")
-		}
-	}
-
-	// don't support negative values
-	if cfg.Flush.Messages < 0 || cfg.Flush.Bytes < 0 || cfg.Flush.Frequency < 0 ||
-		cfg.Retry.Max < 0 || cfg.Retry.Backoff < 0 {
-		return batchCfg, retryCfg, errors.Errorf("invalid sink config, all values must be non-negative")
-	}
-
-	// errors if other batch values are set, but frequency is not
-	if (cfg.Flush.Messages > 0 || cfg.Flush.Bytes > 0) && cfg.Flush.Frequency == 0 {
-		return batchCfg, retryCfg, errors.Errorf("invalid sink config, Flush.Frequency is not set, messages may never be sent")
-	}
-
-	retryCfg.MaxRetries = int(cfg.Retry.Max)
-	retryCfg.InitialBackoff = time.Duration(cfg.Retry.Backoff)
-	retryCfg.MaxBackoff = 30 * time.Second
-	return cfg.Flush, retryCfg, nil
-}
-
-func makeDeprecatedWebhookSink(
+func makeWebhookSinkClient(
 	ctx context.Context,
 	u sinkURL,
 	encodingOpts changefeedbase.EncodingOptions,
 	opts changefeedbase.WebhookSinkOptions,
+	batchCfg sinkBatchConfig,
 	parallelism int,
-	source timeutil.TimeSource,
-	mb metricsRecorderBuilder,
-) (Sink, error) {
-	if u.Scheme != changefeedbase.SinkSchemeWebhookHTTPS {
-		return nil, errors.Errorf(`this sink requires %s`, changefeedbase.SinkSchemeWebhookHTTPS)
+) (SinkClient, error) {
+	err := validateWebhookOpts(u, encodingOpts, opts)
+	if err != nil {
+		return nil, err
 	}
+
 	u.Scheme = strings.TrimPrefix(u.Scheme, `webhook-`)
 
-	switch encodingOpts.Format {
-	case changefeedbase.OptFormatJSON:
-	case changefeedbase.OptFormatCSV:
-	default:
-		return nil, errors.Errorf(`this sink is incompatible with %s=%s`,
-			changefeedbase.OptFormat, encodingOpts.Format)
-	}
-
-	switch encodingOpts.Envelope {
-	case changefeedbase.OptEnvelopeWrapped, changefeedbase.OptEnvelopeBare:
-	default:
-		return nil, errors.Errorf(`this sink is incompatible with %s=%s`,
-			changefeedbase.OptEnvelope, encodingOpts.Envelope)
-	}
-
-	encodingOpts.TopicInValue = true
-
-	if encodingOpts.Envelope != changefeedbase.OptEnvelopeBare {
-		encodingOpts.KeyInValue = true
+	sinkClient := &webhookSinkClient{
+		ctx:        ctx,
+		authHeader: opts.AuthHeader,
+		format:     encodingOpts.Format,
+		batchCfg:   batchCfg,
 	}
 
 	var connTimeout time.Duration
 	if opts.ClientTimeout != nil {
 		connTimeout = *opts.ClientTimeout
 	}
-
-	ctx, cancel := context.WithCancel(ctx)
-
-	sink := &deprecatedWebhookSink{
-		workerCtx:   ctx,
-		authHeader:  opts.AuthHeader,
-		exitWorkers: cancel,
-		parallelism: parallelism,
-		ts:          source,
-		metrics:     mb(requiresResourceAccounting),
-		format:      encodingOpts.Format,
-	}
-
-	var err error
-	sink.batchCfg, sink.retryCfg, err = sink.getWebhookSinkConfig(opts.JSONConfig)
-	if err != nil {
-		return nil, errors.Wrapf(err, "error processing option %s", changefeedbase.OptWebhookSinkConfig)
-	}
-
-	// TODO(yevgeniy): Establish HTTP connection in Dial().
-	sink.client, err = deprecatedMakeWebhookClient(u, connTimeout)
+	sinkClient.client, err = makeWebhookClient(u, connTimeout, parallelism)
 	if err != nil {
 		return nil, err
 	}
@@ -298,17 +98,23 @@ func makeDeprecatedWebhookSink(
 	params.Del(changefeedbase.SinkParamClientCert)
 	params.Del(changefeedbase.SinkParamClientKey)
 	sinkURLParsed.RawQuery = params.Encode()
-	sink.url = sinkURL{URL: sinkURLParsed}
+	sinkClient.url = sinkURL{URL: sinkURLParsed}
 
-	return sink, nil
+	return sinkClient, nil
 }
 
-func deprecatedMakeWebhookClient(u sinkURL, timeout time.Duration) (*httputil.Client, error) {
+func makeWebhookClient(
+	u sinkURL, timeout time.Duration, parallelism int,
+) (*httputil.Client, error) {
 	client := &httputil.Client{
 		Client: &http.Client{
 			Timeout: timeout,
 			Transport: &http.Transport{
-				DialContext: (&net.Dialer{Timeout: timeout}).DialContext,
+				DialContext:         (&net.Dialer{Timeout: timeout}).DialContext,
+				MaxConnsPerHost:     parallelism,
+				MaxIdleConnsPerHost: parallelism,
+				IdleConnTimeout:     time.Minute,
+				ForceAttemptHTTP2:   true,
 			},
 		},
 	}
@@ -370,217 +176,34 @@ func deprecatedMakeWebhookClient(u sinkURL, timeout time.Duration) (*httputil.Cl
 	return client, nil
 }
 
-// defaultWorkerCount() is the number of CPU's on the machine
-func defaultWorkerCount() int {
-	return system.NumCPU()
-}
-
-func (s *deprecatedWebhookSink) Dial() error {
-	s.setupWorkers()
-	return nil
-}
-
-func (s *deprecatedWebhookSink) setupWorkers() {
-	// setup events channels to send to workers and the worker group
-	s.eventsChans = make([]chan []deprecatedMessagePayload, s.parallelism)
-	s.workerGroup = ctxgroup.WithContext(s.workerCtx)
-	s.batchChan = make(chan webhookMessage)
-
-	// an error channel with buffer for the first error.
-	s.errChan = make(chan error, 1)
-
-	// flushDone notified when flush completes.
-	s.flushDone = make(chan struct{})
-
-	s.workerGroup.GoCtx(func(ctx context.Context) error {
-		s.batchWorker()
-		return nil
-	})
-	for i := 0; i < s.parallelism; i++ {
-		s.eventsChans[i] = make(chan []deprecatedMessagePayload)
-		j := i
-		s.workerGroup.GoCtx(func(ctx context.Context) error {
-			s.workerLoop(j)
-			return nil
-		})
-	}
-}
-
-func (s *deprecatedWebhookSink) shouldSendBatch(b batch) bool {
-	// similar to sarama, send batch if:
-	// everything is zero (default)
-	// any one of the conditions are met UNLESS the condition is zero which means never batch
-	switch {
-	// all zero values should batch every time, otherwise batch will wait forever
-	case s.batchCfg.Messages == 0 && s.batchCfg.Bytes == 0 && s.batchCfg.Frequency == 0:
-		return true
-	// messages threshold has been reached
-	case s.batchCfg.Messages > 0 && len(b.buffer) >= s.batchCfg.Messages:
-		return true
-	// bytes threshold has been reached
-	case s.batchCfg.Bytes > 0 && b.bufferBytes >= s.batchCfg.Bytes:
-		return true
-	default:
-		return false
-	}
-}
-
-func (s *deprecatedWebhookSink) splitAndSendBatch(batch []deprecatedMessagePayload) error {
-	workerBatches := make([][]deprecatedMessagePayload, s.parallelism)
-	for _, msg := range batch {
-		// split batch into per-worker batches
-		i := s.workerIndex(msg.key)
-		workerBatches[i] = append(workerBatches[i], msg)
-	}
-	for i, workerBatch := range workerBatches {
-		// don't send empty batches
-		if len(workerBatch) > 0 {
-			select {
-			case <-s.workerCtx.Done():
-				return s.workerCtx.Err()
-			case s.eventsChans[i] <- workerBatch:
-			}
-		}
-	}
-	return nil
-}
-
-// flushWorkers sends flush request to each worker and waits for each one to
-// acknowledge.
-func (s *deprecatedWebhookSink) flushWorkers(done chan struct{}) error {
-	for i := 0; i < len(s.eventsChans); i++ {
-		// Ability to write a nil message to events channel indicates that
-		// the worker has processed all other messages.
-		select {
-		case <-s.workerCtx.Done():
-			return s.workerCtx.Err()
-		case s.eventsChans[i] <- nil:
-		}
-	}
-
-	select {
-	case <-s.workerCtx.Done():
-		return s.workerCtx.Err()
-	case done <- struct{}{}:
-		return nil
-	}
-}
-
-// batchWorker ingests messages from EmitRow into a batch and splits them into
-// per-worker batches to be sent separately
-func (s *deprecatedWebhookSink) batchWorker() {
-	var batchTracker batch
-	batchTimer := s.ts.NewTimer()
-	defer batchTimer.Stop()
-
-	for {
-		select {
-		case <-s.workerCtx.Done():
-			return
-		case msg := <-s.batchChan:
-			flushRequested := msg.flushDone != nil
-
-			if !flushRequested {
-				batchTracker.addToBuffer(msg.payload)
-			}
-
-			if s.shouldSendBatch(batchTracker) || flushRequested {
-				if err := s.splitAndSendBatch(batchTracker.buffer); err != nil {
-					s.exitWorkersWithError(err)
-					return
-				}
-				batchTracker.reset()
-
-				if flushRequested {
-					if err := s.flushWorkers(*msg.flushDone); err != nil {
-						s.exitWorkersWithError(err)
-						return
-					}
-				}
-			} else {
-				if len(batchTracker.buffer) == 1 && time.Duration(s.batchCfg.Frequency) > 0 {
-					// only start timer when first message appears
-					batchTimer.Reset(time.Duration(s.batchCfg.Frequency))
-				}
-			}
-		// check the channel for time expiry. the batch should have at least one
-		// message in it. If it doesn't, the timer has been carried over from a
-		// previous batch, and the new batch will be empty so it won't send. If
-		// the new batch has at least one element, the timer will be reset so it'll
-		// be updated.
-		case <-batchTimer.Ch():
-			batchTimer.MarkRead()
-			if len(batchTracker.buffer) > 0 {
-				if err := s.splitAndSendBatch(batchTracker.buffer); err != nil {
-					s.exitWorkersWithError(err)
-					return
-				}
-				batchTracker.reset()
-			}
-		}
-	}
-}
-
-func (s *deprecatedWebhookSink) workerLoop(workerIndex int) {
-	for {
-		select {
-		case <-s.workerCtx.Done():
-			return
-		case msgs := <-s.eventsChans[workerIndex]:
-			if msgs == nil {
-				// It's a flush request: if we read it, it means all outstanding
-				// requests for this worker have been completed.
-				continue
-			}
-
-			var encoded encodedPayload
-			var err error
-			switch s.format {
-			case changefeedbase.OptFormatJSON:
-				encoded, err = encodePayloadJSONWebhook(msgs)
-			case changefeedbase.OptFormatCSV:
-				encoded, err = encodePayloadCSVWebhook(msgs)
-			}
-			if err != nil {
-				s.exitWorkersWithError(err)
-				return
-			}
-			if err := s.sendMessageWithRetries(s.workerCtx, encoded.data); err != nil {
-				s.exitWorkersWithError(err)
-				return
-			}
-			encoded.alloc.Release(s.workerCtx)
-			s.metrics.recordEmittedBatch(
-				encoded.emitTime, len(msgs), encoded.mvcc, len(encoded.data), sinkDoesNotCompress)
-		}
-	}
-}
-
-func (s *deprecatedWebhookSink) sendMessageWithRetries(ctx context.Context, reqBody []byte) error {
-	requestFunc := func() error {
-		return s.sendMessage(ctx, reqBody)
-	}
-	return retry.WithMaxAttempts(ctx, s.retryCfg, s.retryCfg.MaxRetries+1, requestFunc)
-}
-
-func (s *deprecatedWebhookSink) sendMessage(ctx context.Context, reqBody []byte) error {
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, s.url.String(), bytes.NewReader(reqBody))
+func (sc *webhookSinkClient) makePayloadForBytes(body []byte) (SinkPayload, error) {
+	req, err := http.NewRequestWithContext(sc.ctx, http.MethodPost, sc.url.String(), bytes.NewReader(body))
 	if err != nil {
-		return err
+		return nil, err
 	}
-	switch s.format {
+	switch sc.format {
 	case changefeedbase.OptFormatJSON:
 		req.Header.Set("Content-Type", applicationTypeJSON)
 	case changefeedbase.OptFormatCSV:
 		req.Header.Set("Content-Type", applicationTypeCSV)
 	}
 
-	if s.authHeader != "" {
-		req.Header.Set(authorizationHeader, s.authHeader)
+	if sc.authHeader != "" {
+		req.Header.Set(authorizationHeader, sc.authHeader)
 	}
 
-	var res *http.Response
-	res, err = s.client.Do(req)
+	return req, nil
+}
+
+// MakeResolvedPayload implements the SinkClient interface
+func (sc *webhookSinkClient) MakeResolvedPayload(body []byte, topic string) (SinkPayload, error) {
+	return sc.makePayloadForBytes(body)
+}
+
+// Flush implements the SinkClient interface
+func (sc *webhookSinkClient) Flush(ctx context.Context, batch SinkPayload) error {
+	req := batch.(*http.Request)
+	res, err := sc.client.Do(req)
 	if err != nil {
 		return err
 	}
@@ -596,131 +219,148 @@ func (s *deprecatedWebhookSink) sendMessage(ctx context.Context, reqBody []byte)
 	return nil
 }
 
-// workerIndex assigns rows each to a worker goroutine based on the hash of its
-// primary key. This is to ensure that each message with the same key gets
-// deterministically assigned to the same worker. Since we have a channel per
-// worker, we can ensure per-worker ordering and therefore guarantee per-key
-// ordering.
-func (s *deprecatedWebhookSink) workerIndex(key []byte) uint32 {
-	return crc32.ChecksumIEEE(key) % uint32(s.parallelism)
+// Close implements the SinkClient interface
+func (sc *webhookSinkClient) Close() error {
+	sc.client.CloseIdleConnections()
+	return nil
 }
 
-// exitWorkersWithError saves the first error message encountered by webhook
-// workers,
-// and requests all workers to terminate.
-func (s *deprecatedWebhookSink) exitWorkersWithError(err error) {
-	// errChan has buffer size 1, first error will be saved to the buffer and
-	// subsequent errors will be ignored
-	select {
-	case s.errChan <- err:
-		s.exitWorkers()
+func validateWebhookOpts(
+	u sinkURL, encodingOpts changefeedbase.EncodingOptions, opts changefeedbase.WebhookSinkOptions,
+) error {
+	if u.Scheme != changefeedbase.SinkSchemeWebhookHTTPS {
+		return errors.Errorf(`this sink requires %s`, changefeedbase.SinkSchemeWebhookHTTPS)
+	}
+
+	switch encodingOpts.Format {
+	case changefeedbase.OptFormatJSON:
+	case changefeedbase.OptFormatCSV:
 	default:
+		return errors.Errorf(`this sink is incompatible with %s=%s`,
+			changefeedbase.OptFormat, encodingOpts.Format)
+	}
+
+	switch encodingOpts.Envelope {
+	case changefeedbase.OptEnvelopeWrapped, changefeedbase.OptEnvelopeBare:
+	default:
+		return errors.Errorf(`this sink is incompatible with %s=%s`,
+			changefeedbase.OptEnvelope, encodingOpts.Envelope)
+	}
+
+	encodingOpts.TopicInValue = true
+
+	if encodingOpts.Envelope != changefeedbase.OptEnvelopeBare {
+		encodingOpts.KeyInValue = true
+	}
+
+	return nil
+}
+
+type webhookCSVBuffer struct {
+	bytes        []byte
+	messageCount int
+	sc           *webhookSinkClient
+}
+
+var _ BatchBuffer = (*webhookCSVBuffer)(nil)
+
+// Append implements the BatchBuffer interface
+func (cb *webhookCSVBuffer) Append(key []byte, value []byte) {
+	cb.bytes = append(cb.bytes, value...)
+	cb.messageCount += 1
+}
+
+// ShouldFlush implements the BatchBuffer interface
+func (cb *webhookCSVBuffer) ShouldFlush() bool {
+	return shouldFlushBatch(len(cb.bytes), cb.messageCount, cb.sc.batchCfg)
+}
+
+// Close implements the BatchBuffer interface
+func (cb *webhookCSVBuffer) Close() (SinkPayload, error) {
+	return cb.sc.makePayloadForBytes(cb.bytes)
+}
+
+type webhookJSONBuffer struct {
+	messages [][]byte
+	numBytes int
+	sc       *webhookSinkClient
+}
+
+var _ BatchBuffer = (*webhookJSONBuffer)(nil)
+
+// Append implements the BatchBuffer interface
+func (jb *webhookJSONBuffer) Append(key []byte, value []byte) {
+	jb.messages = append(jb.messages, value)
+	jb.numBytes += len(value)
+}
+
+// ShouldFlush implements the BatchBuffer interface
+func (jb *webhookJSONBuffer) ShouldFlush() bool {
+	return shouldFlushBatch(jb.numBytes, len(jb.messages), jb.sc.batchCfg)
+}
+
+// Close implements the BatchBuffer interface
+func (jb *webhookJSONBuffer) Close() (SinkPayload, error) {
+	var buffer bytes.Buffer
+	prefix := "{\"payload\":["
+	suffix := fmt.Sprintf("],\"length\":%d}", len(jb.messages))
+
+	// Grow all at once to avoid reallocations
+	buffer.Grow(len(prefix) + jb.numBytes /* msgs */ + len(jb.messages) /* commas */ + len(suffix))
+
+	buffer.WriteString(prefix)
+	for i, msg := range jb.messages {
+		if i != 0 {
+			buffer.WriteByte(',')
+		}
+		buffer.Write(msg)
+	}
+	buffer.WriteString(suffix)
+	return jb.sc.makePayloadForBytes(buffer.Bytes())
+}
+
+// MakeBatchBuffer implements the SinkClient interface
+func (sc *webhookSinkClient) MakeBatchBuffer(topic string) BatchBuffer {
+	if sc.format == changefeedbase.OptFormatCSV {
+		return &webhookCSVBuffer{sc: sc}
+	} else {
+		return &webhookJSONBuffer{
+			sc:       sc,
+			messages: make([][]byte, 0, sc.batchCfg.Messages),
+		}
 	}
 }
 
-// sinkError checks to see if any errors occurred inside workers go routines.
-func (s *deprecatedWebhookSink) sinkError() error {
-	select {
-	case err := <-s.errChan:
-		return err
-	default:
-		return nil
-	}
-}
-
-func (s *deprecatedWebhookSink) EmitRow(
+func makeWebhookSink(
 	ctx context.Context,
-	topic TopicDescriptor,
-	key, value []byte,
-	updated, mvcc hlc.Timestamp,
-	alloc kvevent.Alloc,
-) error {
-	select {
-	// check the webhook sink context in case workers have been terminated
-	case <-s.workerCtx.Done():
-		// check again for error in case it triggered since last check
-		// will return more verbose error instead of "context canceled"
-		return errors.CombineErrors(s.workerCtx.Err(), s.sinkError())
-	case <-ctx.Done():
-		return ctx.Err()
-	case err := <-s.errChan:
-		return err
-	case s.batchChan <- webhookMessage{
-		payload: deprecatedMessagePayload{
-			key:      key,
-			val:      value,
-			alloc:    alloc,
-			emitTime: timeutil.Now(),
-			mvcc:     mvcc,
-		}}:
-		s.metrics.recordMessageSize(int64(len(key) + len(value)))
-	}
-	return nil
-}
-
-func (s *deprecatedWebhookSink) EmitResolvedTimestamp(
-	ctx context.Context, encoder Encoder, resolved hlc.Timestamp,
-) error {
-	defer s.metrics.recordResolvedCallback()()
-
-	payload, err := encoder.EncodeResolvedTimestamp(ctx, "", resolved)
+	u sinkURL,
+	encodingOpts changefeedbase.EncodingOptions,
+	opts changefeedbase.WebhookSinkOptions,
+	parallelism int,
+	pacerFactory func() *admission.Pacer,
+	source timeutil.TimeSource,
+	mb metricsRecorderBuilder,
+) (Sink, error) {
+	batchCfg, retryOpts, err := getSinkConfigFromJson(opts.JSONConfig, sinkJSONConfig{})
 	if err != nil {
-		return err
+		return nil, err
 	}
 
-	select {
-	// check the webhook sink context in case workers have been terminated
-	case <-s.workerCtx.Done():
-		return s.workerCtx.Err()
-	// non-blocking check for error, restart changefeed if encountered
-	case <-s.errChan:
-		return err
-	default:
+	sinkClient, err := makeWebhookSinkClient(ctx, u, encodingOpts, opts, batchCfg, parallelism)
+	if err != nil {
+		return nil, err
 	}
 
-	// do worker logic directly here instead (there's no point using workers for
-	// resolved timestamps since there are no keys and everything must be
-	// in order)
-	if err := s.sendMessageWithRetries(ctx, payload); err != nil {
-		s.exitWorkersWithError(err)
-		return err
-	}
-
-	return nil
-}
-
-func (s *deprecatedWebhookSink) Flush(ctx context.Context) error {
-	s.metrics.recordFlushRequestCallback()()
-
-	// Send flush request.
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case err := <-s.errChan:
-		return err
-	case s.batchChan <- webhookMessage{flushDone: &s.flushDone}:
-	}
-
-	// Wait for flush completion -- or an error.
-	select {
-	case <-ctx.Done():
-		return ctx.Err()
-	case err := <-s.errChan:
-		return err
-	case <-s.flushDone:
-		return s.sinkError()
-	}
-}
-
-func (s *deprecatedWebhookSink) Close() error {
-	s.exitWorkers()
-	// ignore errors here since we're closing the sink anyway
-	_ = s.workerGroup.Wait()
-	close(s.batchChan)
-	close(s.errChan)
-	for _, eventsChan := range s.eventsChans {
-		close(eventsChan)
-	}
-	s.client.CloseIdleConnections()
-	return nil
+	return makeBatchingSink(
+		ctx,
+		sinkTypeWebhook,
+		sinkClient,
+		time.Duration(batchCfg.Frequency),
+		retryOpts,
+		parallelism,
+		nil,
+		pacerFactory,
+		source,
+		mb(requiresResourceAccounting),
+	), nil
 }
